// salarasa-app/src/db/schema.ts
export type BlockType =
  | 'splash_cover'
  | 'hero'
  | 'subject_biodata'
  | 'countdown_event'
  | 'dresscode'
  | 'gift'
  | 'rsvp'
  | 'guestbook'
  | 'thank_you'
  | 'footer'